Implemented CachedPropertyEnricher
