+++ src/Serilog.Enrichers.Environment/Enrichers/CachedPropertyEnricher.cs
@@ -1,0 +1,46 @@
+// Copyright 2013-2022 Serilog Contributors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using Serilog.Core;
+using Serilog.Events;
+
+namespace Serilog.Enrichers
+{
+    public abstract class CachedPropertyEnricher: ILogEventEnricher
+    {
+        private LogEventProperty _cachedProperty { get; set; }
+
+        /// <summary>
+        /// Enrich the log event.
+        /// </summary>
+        /// <param name="logEvent">The log event to enrich.</param>
+        /// <param name="propertyFactory">Factory for creating new properties to add to the event.</param>
+        public void Enrich(LogEvent logEvent, ILogEventPropertyFactory propertyFactory)
+        {
+            logEvent.AddPropertyIfAbsent(GetLogEventProperty(propertyFactory));
+        }
+
+        private LogEventProperty GetLogEventProperty(ILogEventPropertyFactory propertyFactory)
+        {
+            // Don't care about thread-safety, in the worst case the field gets overwritten and one
+            // property will be GCed
+            if (_cachedProperty == null)
+                _cachedProperty = CreateProperty(propertyFactory);
+
+            return _cachedProperty;
+        }
+        
+        protected abstract LogEventProperty CreateProperty(ILogEventPropertyFactory propertyFactory);
+    }
+}
