+++ src/Serilog.Enrichers.Environment/Enrichers/MachineNameEnricher.cs
@@ -1,4 +1,4 @@
-﻿// Copyright 2013-2018 Serilog Contributors
+﻿// Copyright 2013-2022 Serilog Contributors
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -26,38 +26,16 @@ namespace Serilog.Enrichers
     /// <summary>
     /// Enriches log events with a MachineName property containing <see cref="Environment.MachineName"/>.
     /// </summary>
-    public class MachineNameEnricher : ILogEventEnricher
+    public class MachineNameEnricher : CachedPropertyEnricher
     {
-        LogEventProperty _cachedProperty;
-
         /// <summary>
         /// The property name added to enriched log events.
         /// </summary>
         public const string MachineNamePropertyName = "MachineName";
 
-        /// <summary>
-        /// Enrich the log event.
-        /// </summary>
-        /// <param name="logEvent">The log event to enrich.</param>
-        /// <param name="propertyFactory">Factory for creating new properties to add to the event.</param>
-        public void Enrich(LogEvent logEvent, ILogEventPropertyFactory propertyFactory)
-        {
-            logEvent.AddPropertyIfAbsent(GetLogEventProperty(propertyFactory));
-        }
-
-        private LogEventProperty GetLogEventProperty(ILogEventPropertyFactory propertyFactory)
-        {
-            // Don't care about thread-safety, in the worst case the field gets overwritten and one
-            // property will be GCed
-            if (_cachedProperty == null)
-                _cachedProperty = CreateProperty(propertyFactory);
-
-            return _cachedProperty;
-        }
-
         // Qualify as uncommon-path
         [MethodImpl(MethodImplOptions.NoInlining)]
-        private static LogEventProperty CreateProperty(ILogEventPropertyFactory propertyFactory)
+        protected override LogEventProperty CreateProperty(ILogEventPropertyFactory propertyFactory)
         {
 #if NETSTANDARD1_3
             var machineName = Dns.GetHostName();
